+++ library/spoon/rest/client.php
@@ -75,7 +75,7 @@ class SpoonRESTClient
 		if(!function_exists('curl_init')) throw new SpoonFileException('This method requires cURL (http://php.net/curl), it seems like the extension isn\'t installed.');
 
 		// init var
-		$allowedMethods = array('GET', 'POST');
+		$allowedMethods = array('GET', 'POST', 'DELETE');
 
 		// redefine
 		$url = (string) $url;